requirements.txt
streamlit
pandas
google-generativeai
python-dotenv
fastapi
uvicorn
pydantic